fix: some forms still require validation even after the module is switched off in integration form
